requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.10.0
python-jose[cryptography]==3.3.0
email-validator
